refactor(sample): Move terminus related stuff into seperate health module
