+++ sample/000-dogs-app/src/main.ts
@@ -3,6 +3,6 @@ import { ApplicationModule } from './app.module';
 
 async function bootstrap() {
   const app = await NestFactory.create(ApplicationModule);
-  await app.listen(3001);
+  await app.listen(3000);
 }
 bootstrap();
